+++ CHANGELOG.md
@@ -10,6 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `basketAddItem` mutation honours stockflag and calls may give you back an
   error
+- Method `OxidEsales\GraphQL\Storefront\Basket\Service\Basket::publicBasketsByOwnerNameOrEmail` now returns an array of `OxidEsales\GraphQL\Storefront\Basket\DataType\PublicBasket`
+- `baskets(owner String)` now returns an array of `OxidEsales\GraphQL\Storefront\Basket\DataType\PublicBasket` in order to not expose address or payment information on a public basket
+
+### Added
+- Classes
+  - `OxidEsales\GraphQL\Storefront\Basket\DataType\PublicBasket`
+  - `OxidEsales\GraphQL\Storefront\Basket\Service\PublicBasketRelationService`
 
 ## [1.0.0] - 2021-05-28